Add test to check location.js object properties
This PR adds a test to try to help ensure that location files are setup properly.

If a location file doesn't have a valid email address for the email or the receivingEmail properties the test will fail.

This will help catch possible entry errors prior to any code getting pushed out.

This also adds the jest-expect-message which allows adding a custom message to an expect case.  This useful as we can indicate the name of the file missing the property.